chore(completion): make nvim-cmp snippets not cry on tabs
Before:
If you're in middle of a snippet, and you've indeed found something you
want to use for that "code location" in the snippet, you'd normally
switch to next "code location" (if present) within the snippet, but it'd
not let you because Tab/S-Tab are tied to moving to next/previous entry
in the completion list, hence they break the whole point of using a
snippet!

After:
We just disable the bindings for Tab/S-Tab, and avoid this problem
altogether. If we do need to go to next/previous entry in the completion
list, we'll basically use the Ctrl-N/Ctrl-P (which are the defaults)